Big Update

+++ core/exception/exception.h
@@ -9,16 +9,16 @@ class Exception : public std::exception
 {
 	private:
 		char* msg = NULL;
-		unsigned short code = 0;
+		unsigned int code = 0;
 
 	public:
-		Exception(unsigned short code, const char* msg) throw();
+		Exception(unsigned int code, const char* msg) throw();
 		~Exception();
 
 		virtual const char* what() const throw();
 
 		const char* getMsg(); 
-		const unsigned short getCode();
+		const unsigned int getCode();
 };
 
 #endif
